Implement two matching heuristics.
(1) Implement sift4 string distance algorithm.
(2) Implement heuristic that if the original MM algorithm does not
match the path. It will match the last token of the path instead.

It turns out that using the last token of the path works pretty well.
So, instead of using a more complicated matching algorithm, we are using
(2).